Expose Job parameters in JdbcSearchableJobExecutionDao
See https://jira.springsource.org/browse/BATCHADM-176

Must be applied after
https://github.com/spring-projects/spring-batch/pull/231

As part of this, I had to update the dependency on spring-batch to
3.0.0-BUILD_SNAPSHOT.

Because of this dependency change, I had to add two
methods to JobRepositorySupport.  It looks like a stub class, so I
left the two methods implemented as stubs.