Remove spring-batch-integration
As part of Spring Batch 3.0, the spring-batch-integration module was
moved over from Spring Batch Admin to Spring Batch and upgraded to
support Spring Integration 4.  This commit removes the old version of
the module.  The only impact on this (beyond the Spring Integraiton 4
requirement) is the maven location change which has been available for
almost a year now.

This commit addresses BATCHADM-205