Updated javadoc to pass doclint
JDK 1.8 turned on doclint, validating javadoc and subsequently failing
builds when generating javadoc with invalid markup.  This commit
addresses the javadoc errors in this project that cause javadoc
generation to fail.  There are still warnings that need to be addressed,
but this project should now build with JDK 1.8 when generating javadoc.

BATCHADM-216