Build fails on windows.
The method URI#normalize fails on windows because of the colon ':' char
contained in the path (C:\path\to\file). I propose to use
File#getCanonicalPath instead. Plus a test fails (ConfigurationViewTests) because of the
${batch.files.upload-dir:#{systemProperties['java.io.tmpdir']}} parameter
in file-context.xml. Again the ':' char could have been an issue, but
anyway batch.files.upload-dir definition was missing for this test.